Fixed the scaladoc/rangepos problem for good
the problem was that comments were "delayed" because the terms they were
appearing above were not supposed to take documentation (ex: cases in a
match statement) => the comments were added to later terms (that would
normally be documented, such as defs or vals). To prevent that, we now
check that only whitespace is separating the comment and the term we're
attaching the comment to -- otherwise we simply drop the comment, as it
was attached to a node that does not need documentation.