Step 2: Fixing the _this type
Okay, so here's the brief motivation of _this:
```
class C {
  final def tailCall(other: C): Nothing = other.tailCall
}
```
Now you want to jump to the same label for the tailcall, but change
the this you're referring to. It turns out this is what the backend
has been doing whenever it found a call to a label with the _this
parameter (ASTORE_0). So there's no need to have the ThisSubstituter,
which was the reason of the latest batch of crashes.

Now on to the last problem: getting the _this functionality back.