SI-4767 Inline @inline final trait methods
A trait consists of an interface and an implementation class. From a
JVM perspective, a class that extends a trait will actually implement
the interface and unless it overrides the member, will get a forwarder
to the implementation class:

for:
```
  trait Foo { @inline final def foo(x: Int) = println(x) }
  class Baz extends Foo
```

Baz is compled down to:
```
  class Baz extends Object with Foo {
   @inline final def foo(x: Int): Unit = Foo$class.foo(FooImpl.this, x)
   ...
  }
```

Now, say we have:
```
  def bar(y: Foo) = y.foo(7)
```

The type of `Foo` is `interface Foo`. Thus there's no way for the
inliner to load the bytecode for foo. The first part of this patch
addresses this problem, by transforming the call to `y.foo(7)` to
`Foo$class.foo(y, 7)` if the foo method is final and @inline marked.
This first stage of the inlining is done in GenICode.scala and is not
necessary for inlining if we explicitly have `Baz` insted of `Foo` as
the type of x. (the inliner will inline the forwarder)

The second part is actually inlining the static call to `Foo$class.foo`.
Now, the trick is that the inliner actually expects the calls it
inlines to have an instance on which they're run. I relaxed the inline
conditions to include static calls and changed the inliner to generate
correct code for static calls. (the VerifyError was caused by this)

Finally, added a test with Greg's instrumented partest addon, ant it's
awesome! Greg promised to add more tests later.

Review by @mgarciaEPFL and @gkossakowski.