SI-6248 Correct bytecode for box/unbox sequences
Even for semantically incorrect code, the closure elimination phase
continues to generate correct java bytecode.

The sequence in question:
```
   LOAD_LOCAL v1  // <= loads v1: Long onto the stack
   BOX LONG       // <= takes a Long, returns an Object
   UNBOX INT      // <= takes an Object, returns an Int
```

From a Scala perspective, this can't happen. But with aggresive inline,
there is a way to produce this code. It won't ever be executed directly
and if expressly called, should result in a runtime exception. But with
the closure elimination (better called copy propagation phase) this
code will result in an incorrect:
```
   LOAD_LOCAL v1
   BOX LONG
   LOAD_LOCAL v1  // <= an Int should be on the stack, but we get a
                  // 2-slot long
```