Hack for a reification bug
That is SI-5875 (not in title because we just reference it)

Perform eta expansion on methods before plugging them into macros:
```
val free$z1 = build.newFreeTerm("z", z, METHOD | TRIEDCOOKING, "defined by fun in t5875-a.scala:8:9");
```
now becomes:
```
val free$z1 = build.newFreeTerm("z", (() => z()), METHOD | TRIEDCOOKING, "defined by fun in t5875-a.scala:8:9");
```
The catch is that methods can't be passed as values while functions
can. So each time we encounter a methods we create a function node.

This doesn't fix the problem in t5875-a.scala. But it's a start for
the rest of the bugfix.